Move StepNav model inline
This model is a private class that's only used inside the stepnav
helper. Moving it here will make it easier to understand what the
difference is between all the StepNav code.
