+++ lib/govuk_publishing_components.rb
@@ -5,7 +5,6 @@ require "govuk_publishing_components/presenters/navigation_type"
 require "govuk_publishing_components/presenters/related_navigation_helper"
 require "govuk_publishing_components/presenters/step_by_step_nav_helper"
 require "govuk_publishing_components/presenters/step_nav_helper"
-require "govuk_publishing_components/presenters/step_nav"
 
 require "govuk_publishing_components/app_helpers/taxon_breadcrumbs"
